pyproject.toml
[tool.poetry]
name = "pdf-bot"
version = "0.1.0"
description = ""
authors = ["alexander.evseev <alexandr.evseev@balance-pl.ru>"]
readme = "README.md"
packages = [{include = "pdf_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.2"
pydantic = "^1.10.7"
requests = "^2.31.0"
"pdfminer.six" = {extras = ["image"], version = "^20221105"}
pdftotext = "^2.2.2"
isort = "^5.12.0"
flake8 = "^6.1.0"
black = "^23.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
